Hotfix for Player Dead Body Movement
Souls should not be moving their deadbodies around the map.  Updated Character View Offset to use 'movementId' instead of Character Instance ID

+++ Necromancy.Server/Chat/Command/Commands/Migrated/Revive.cs
@@ -27,7 +27,7 @@ namespace Necromancy.Server.Chat.Command.Commands
                 res.WriteInt32(0); // 0 = sucess to revive, 1 = failed to revive
                 client.Character.soulFormState -= 1;
                 client.Character.currentHp = client.Character.maxHp;
-                client.Character.movementId = (int)client.Character.InstanceId;
+                client.Character.movementId = client.Character.InstanceId;
                 Router.Send(client, (ushort) AreaPacketId.recv_raisescale_request_revive_r, res, ServerType.Area);
 
                 IBuffer res2 = BufferProvider.Provide();
