+++ Necromancy.Server/Packet/Msg/send_chara_select.cs
@@ -62,7 +62,7 @@ namespace Necromancy.Server.Packet.Msg
             Server.Instances.AssignInstance(deadBody);
             character.DeadBodyInstanceId = deadBody.InstanceId;
             deadBody.CharacterInstanceId = character.InstanceId;
-            character.movementId = (int)character.InstanceId;
+            character.movementId = character.InstanceId;
             Logger.Debug($"Dead Body Instance ID {deadBody.InstanceId}   |  Character Instance ID {character.InstanceId}");
             deadBody.CharaName = character.Name;
             deadBody.MapId = character.MapId;
